feat(c-api) Move `function` into its own module.
