+++ lib/c-api/src/wasm_c_api/externals/mod.rs
@@ -1,3 +1,4 @@
+pub mod function;
 pub mod global;
 pub mod memory;
 pub mod table;
